pyproject.toml
[project]
name = "lab-9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "hf-xet>=1.2.0",
    "kokoro>=0.9.4",
    "openai-whisper>=20250625",
    "pip>=25.3",
    "pydantic-ai>=1.32.0",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
